LPS-45452 Add support for localized values in BaseServiceSettings
Note that I've made _fallbackSettings private. This is to avoid overriders' access to it as it is the lowest level settings object. All overriders should use the high level settings objects (as of now typedSettings and localizedSettings) which in the future can be expanded if more specialized functionality is needed.